Split out test into their own files
